pyproject.toml
[tool.poetry]
name = "resume"
version = "2.0.0"
description = ""
authors = ["Jack Greenberg <j.lester.greenberg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
jinja2 = "^2.11.2"
click = "^7.1.2"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
build = "resume.app:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
